feat: add GitHub Actions CI/CD and Changesets release pipeline
- Add CI workflow running tests on push/PR with Node 18+20 matrix
- Add release workflow using @changesets/action to open versioning PRs
  and publish to npm on merge
- Initialize Changesets with public access and master base branch
- Add @changesets/cli to devDependencies
- Fix build/watch scripts to use && for fail-fast behavior
- Remove obsolete upver script (superseded by Changesets)
- Pin actions/checkout, actions/setup-node, changesets/action to immutable
  commit SHAs instead of floating tags
- Add persist-credentials: false to checkout steps
- Enable Corepack before yarn invocations so Yarn 4.5.0 is activated
- Restore packageManager field in package.json (needed by Corepack)
- Document Yarn 4 Berry + GitHub Actions + Changesets pattern

Secrets required in repo settings: NPM_TOKEN
