+++ package.json
@@ -28,15 +28,14 @@
     "src"
   ],
   "scripts": {
-    "build": "yarn clean ; yarn tsc -p . ; yarn tsc -p src",
+    "build": "yarn clean && yarn tsc -p . && yarn tsc -p src",
     "clean": "yarn rimraf ./lib ./build",
     "prepublish": "yarn build",
     "test": "yarn build && yarn eslint && yarn jest --verbose test/*.ts",
     "thrift": "thrift --gen js:node parquet.thrift && thrift --gen js:ts parquet.thrift",
     "tsgen": "thrift-typescript --target apache --rootDir . --sourceDir . --outDir codegen parquet.thrift",
     "tsgencore": "thrift-typescript --target thrift-server --rootDir . --sourceDir . --outDir codegen parquet.thrift",
-    "upver": "npm version patch && git push --follow-tags",
-    "watch": "yarn clean ; yarn tsc --watch"
+    "watch": "yarn clean && yarn tsc --watch"
   },
   "engines": {
     "node": ">=18"
@@ -55,7 +54,6 @@
   },
   "devDependencies": {
     "@creditkarma/thrift-typescript": "^3.7.2",
-    "@duckdb/node-api": "^1.2.0",
     "@types/chai": "^4.1.7",
     "@types/debug": "^4.1.4",
     "@types/jest": "^24.0.17",
@@ -82,7 +80,8 @@
     "test-jest": "^1.0.1",
     "ts-jest": "^27.0.7",
     "ts-node": "^10.3.0",
-    "typescript": "^4.4.4"
+    "typescript": "^4.4.4",
+    "@changesets/cli": "^2.27.0"
   },
   "peerDependencies": {
     "@types/node": "*",
